NetEmbryo library: one step more in clamming it docs: TODO new man page as soon as possible sdp2: sdp_session_descr declaration according to sdp_session_descr.c  corrected.

+++ include/fenice/sdp2.h
@@ -48,7 +48,7 @@
 
 gint sdp_session_id(char *, size_t);
 gint sdp_get_version(ResourceDescr *, char *, size_t);
-int sdp_session_descr(resource_name, int, char *, uint32);
+int sdp_session_descr(resource_name, int, char *, size_t);
 int sdp_media_descr(ResourceDescr *, MediaDescrList, char *, uint32);
 
 #endif /*SDP2_H_*/
